requirements.txt
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/avx2
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/generic
backcall==0.1.0
bleach==2.1.3
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.2
entrypoints==0.2.3
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.16.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
pyzmq==17.1.2
qtconsole==4.3.1
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch-gpu==1.0.0
tornado==5.1.1
tqdm==4.24.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
